+++ tests/test_video.py
@@ -70,6 +70,11 @@ def test_video_thumbnail(mp4_file):
     assert isinstance(thumb, Image.Image)
 
 
+def test_video_thumbnail_fail(ogv_file):
+    thumb = idk.video_thumbnail(ogv_file)
+    assert thumb is None
+
+
 def test_video_mp7sig_extract(mp4_file):
     sig = idk.video_mp7sig_extract(mp4_file)
     assert sig[-32:].hex() == "9ef43526febb8d3e674975584ad6812ccc144cba28b3e134cd173888449cf51e"